data insertion into db-tables is fixed - yay!

+++ src/main/java/org/example/model/Krankenhaus.java
@@ -32,7 +32,7 @@ public class Krankenhaus {
                 '}';
     }
 
-    public long getKrankehausId () {
+    public int getKrankehausId () {
         return krankehausId;
     }
